Needed to downgraded to torch 1.7.1. due to CUDA error with new GPUs.

+++ train/loss_functions.py
@@ -21,7 +21,7 @@ class XEDiceLoss(nn.Module):
 #         print(type(tensor(0, dtype=torch.double).cuda()))
         # Cross-entropy loss
         temp_true = torch.where((true == 255), 0, true) # cast 255 to 0 temporarily
-        xe_loss = self.xe(pred, temp_true)
+        xe_loss = self.xe(pred, temp_true).float()
         xe_loss = xe_loss.masked_select(valid_pixel_mask).mean()
 
         # Dice loss
@@ -30,4 +30,4 @@ class XEDiceLoss(nn.Module):
         true = true.masked_select(valid_pixel_mask)
         dice_loss = 1 - (2.0 * torch.sum(pred * true)) / (torch.sum(pred + true) + 1e-7)
 
-        return (0.5 * xe_loss) + (0.5 * dice_loss)
+        return (0.5 * xe_loss.long()) + (0.5 * dice_loss)
